// src/generated/apis/LiturgyFlower/LiturgyFlower.api.ts
import { ContentType, HttpClient, RequestParams } from '../@http-client'
import {
  CommonErrorType,
  LiturgyFlowerErrorMessageType,
  LiturgyFlowerRequestType,
  LiturgyFlowerType,
  PaginatedLiturgyFlowerListType,
} from '../@types/data-contracts'
import { DeepOmitReadOnly } from '../@types/util-types'

/**
 * !DO NOT EDIT THIS FILE
 *
 * 스크립트가 실행될때, 파일을 항상 새로 쓰기 때문에 파일 수정시 작성내용이 제거 될 수 있습니다.
 */

export class LiturgyFlowerApi<
  SecurityDataType = unknown,
> extends HttpClient<SecurityDataType> {
  /**
   * No description
   *
   * @tags liturgy_flower
   * @name LiturgyFlowerList
   * @summary 전례꽃 목록 조회
   * @request GET:/v1/liturgy_flower/
   * @secure
   */

  liturgyFlowerList = (variables?: {
    query?: {
      /** Number of results to return per page. */
      limit?: number
      /** The initial index from which to return the results. */
      offset?: number
    }
    params?: RequestParams
  }) =>
    this.request<PaginatedLiturgyFlowerListType, CommonErrorType>({
      path: `/v1/liturgy_flower/`,
      method: 'GET',
      query: variables?.query,
      secure: true,
      format: 'json',
      ...variables?.params,
    }) /**
   * No description
   *
   * @tags liturgy_flower
   * @name LiturgyFlowerCreate
   * @summary 전례꽃 생성
   * @request POST:/v1/liturgy_flower/
   * @secure
   */

  liturgyFlowerCreate = (variables: {
    data: DeepOmitReadOnly<LiturgyFlowerRequestType>
    params?: RequestParams
  }) =>
    this.request<
      LiturgyFlowerType,
      LiturgyFlowerErrorMessageType | CommonErrorType
    >({
      path: `/v1/liturgy_flower/`,
      method: 'POST',
      body: variables.data,
      secure: true,
      type: ContentType.Json,
      format: 'json',
      ...variables.params,
    }) /**
   * No description
   *
   * @tags liturgy_flower
   * @name LiturgyFlowerRetrieve
   * @summary 전례꽃 상세 조회
   * @request GET:/v1/liturgy_flower/{id}/
   * @secure
   */

  liturgyFlowerRetrieve = (variables: { id: number; params?: RequestParams }) =>
    this.request<LiturgyFlowerType, CommonErrorType>({
      path: `/v1/liturgy_flower/${variables.id}/`,
      method: 'GET',
      secure: true,
      format: 'json',
      ...variables.params,
    }) /**
   * No description
   *
   * @tags liturgy_flower
   * @name LiturgyFlowerUpdate
   * @summary 전례꽃 수정
   * @request PUT:/v1/liturgy_flower/{id}/
   * @secure
   */

  liturgyFlowerUpdate = (variables: {
    id: number
    data: DeepOmitReadOnly<LiturgyFlowerRequestType>
    params?: RequestParams
  }) =>
    this.request<
      LiturgyFlowerType,
      LiturgyFlowerErrorMessageType | CommonErrorType
    >({
      path: `/v1/liturgy_flower/${variables.id}/`,
      method: 'PUT',
      body: variables.data,
      secure: true,
      type: ContentType.Json,
      format: 'json',
      ...variables.params,
    }) /**
   * No description
   *
   * @tags liturgy_flower
   * @name LiturgyFlowerDestroy
   * @summary 전례꽃 삭제
   * @request DELETE:/v1/liturgy_flower/{id}/
   * @secure
   */

  liturgyFlowerDestroy = (variables: { id: number; params?: RequestParams }) =>
    this.request<void, CommonErrorType>({
      path: `/v1/liturgy_flower/${variables.id}/`,
      method: 'DELETE',
      secure: true,
      ...variables.params,
    })
}

//